Local changes to start moving to TypedChainId

+++ src/events_watcher/proposal_handler_watcher.rs
@@ -80,7 +80,7 @@ impl SubstrateEventWatcher for ProposalHandlerWatcher {
                 .webb_config
                 .evm
                 .values()
-                .find(|c| c.chain_id == u64::from(id))
+                .find(|c| c.chain_id == u32::from(id))
                 .and_then(|c| {
                     c.contracts.iter().find(|contract| {
                         matches!(contract, Contract::SignatureBridge(_))
